Extract null_checks and models helpers from session.py
Moves two stand-alone concerns out of session.py:

- dsgrid/ibis/null_checks.py: check_for_nulls. The function issues N+1
  SQL queries on failure (one per column with at least one NULL) and
  is on the Phase 11 perf-fix list; moving it to its own file keeps
  that rewrite scoped to a small surface.

- dsgrid/ibis/models.py: models_to_dataframe and get_type_from_union.
  Both helpers reflect Pydantic model fields into a PySpark-style
  schema; reviewer noted the location was odd. They now live with each
  other and with their TODO comment about Union/None type detection.

Both modules use lazy session imports to break the circular dependency
(session.py imports these modules' siblings during bootstrap). Drops
several now-unused imports from session.py (enum, UnionType, Union,
Type, get_args, get_origin, DSGBaseModel, DSGInvalidField,
handle_column_spaces, is_table_empty).

11 call sites updated across dsgrid/ and tests/. Test status: 175
passed, 5 failed where all 5 failures are the pre-existing
_DuckDBReader.csv WIP, unrelated to this change.
