+++ dsgrid/utils/dataset.py
@@ -51,8 +51,8 @@ from dsgrid.ibis.table_utils import table_to_records
 from dsgrid.ibis.types import is_table_empty, use_duckdb
 from dsgrid.utils.scratch_dir_context import ScratchDirContext
 from dsgrid.ibis.io import persist_table, write_dataframe
+from dsgrid.ibis.null_checks import check_for_nulls
 from dsgrid.ibis.session import (
-    check_for_nulls,
     get_spark_session,
     get_runtime_session,
 )
